(feat) 멀티모듈 적용
