+++ hanghae-cinema/adapter/src/main/java/org/example/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package org.example.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
