+++ hanghae-cinema/adapter/src/main/java/org/example/outbound/persistence/entity/ScreeningScheduleJpaEntity.java
@@ -1,0 +1,29 @@
+package org.example.outbound.persistence.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Entity
+@Table(name = "screening_schedules")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ScreeningScheduleJpaEntity extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "screening_id", nullable = false)
+    private ScreeningJpaEntity screening;
+
+    @Column(nullable = false)
+    private LocalDateTime startTime;
+
+    @Column(nullable = false)
+    private LocalDateTime endTime;
+}
