+++ hanghae-cinema/adapter/src/main/resources/application.yaml
