+++ hanghae-cinema/adapter/src/main/resources/mysql/init/ddl.sql
