+++ hanghae-cinema/application/src/main/java/org/example/port/inbound/ScreeningUseCase.java
@@ -1,0 +1,10 @@
+package org.example.port.inbound;
+
+
+import org.example.port.inbound.dto.MovieResponseDto;
+
+import java.util.List;
+
+public interface ScreeningUseCase {
+    List<MovieResponseDto> getNowShowingMovies();
+}
