+++ hanghae-cinema/domain/src/main/java/org/example/domain/ScreeningSchedule.java
@@ -1,0 +1,19 @@
+package org.example.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+public class ScreeningSchedule {
+    private Long id;
+    private Screening screening;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+    private String createdBy;
+    private String updatedBy;
+}
