--- hanghae-cinema/src/main/java/com/example/hanghaecinema/HanghaeCinemaApplication.java
@@ -1,13 +1,0 @@
-package com.example.hanghaecinema;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HanghaeCinemaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HanghaeCinemaApplication.class, args);
-	}
-
-}
