--- hanghae-cinema/src/main/java/com/example/hanghaecinema/common/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.example.hanghaecinema.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
