--- hanghae-cinema/src/main/java/com/example/hanghaecinema/domain/Reservation.java
@@ -1,34 +1,0 @@
-package com.example.hanghaecinema.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Entity
-@Table(name = "reservations", uniqueConstraints = {
-        @UniqueConstraint(columnNames = {"screening_id", "seat_number"})
-})
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Reservation extends BaseEntity{
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "screening_id", nullable = false)
-    private Screening screening;
-
-    private String userId;
-
-    @Column(nullable = false, length = 5)
-    private String seatNumber;
-
-    @Column(nullable = false)
-    private LocalDateTime reservedAt = LocalDateTime.now();
-
-}
