--- hanghae-cinema/src/main/java/com/example/hanghaecinema/domain/Screening.java
@@ -1,35 +1,0 @@
-package com.example.hanghaecinema.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Entity
-@Table(name = "screenings")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Screening extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "movie_id", nullable = false)
-    private Movie movie;
-
-    @Column(nullable = false, length = 255)
-    private String theater;
-
-    @Column(nullable = false)
-    private LocalDate showDate;
-
-    @OneToMany(mappedBy = "screening", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
-    private List<ScreeningSchedule> schedules = new ArrayList<>();
-
-}
