--- hanghae-cinema/src/main/java/com/example/hanghaecinema/domain/ScreeningSchedule.java
@@ -1,29 +1,0 @@
-package com.example.hanghaecinema.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Entity
-@Table(name = "screening_schedules")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ScreeningSchedule extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "screening_id", nullable = false)
-    private Screening screening;
-
-    @Column(nullable = false)
-    private LocalDateTime startTime;
-
-    @Column(nullable = false)
-    private LocalDateTime endTime;
-}
