--- hanghae-cinema/src/main/java/com/example/hanghaecinema/service/ScreeningService.java
@@ -1,9 +1,0 @@
-package com.example.hanghaecinema.service;
-
-import com.example.hanghaecinema.controller.dto.MovieResponseDto;
-
-import java.util.List;
-
-public interface ScreeningService {
-    List<MovieResponseDto> getNowShowingMovies();
-}
