--- hanghae-cinema/src/test/java/com/example/hanghaecinema/HanghaeCinemaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.hanghaecinema;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HanghaeCinemaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
